code change in Shipper module to fetch getById Shipper

+++ DPUWebProject/src/main/java/com/dpu/controller/ShipperController.java
@@ -134,7 +134,7 @@ public class ShipperController extends MessageProperties {
 	}
 
 	@RequestMapping(value = "/{shipperId}", produces = MediaType.APPLICATION_JSON_VALUE, method = RequestMethod.GET)
-	public Object get(@PathVariable("shipperId") int id) {
+	public Object get(@PathVariable("shipperId") Long id) {
 		logger.info("[get ] : Enter: Id : "+id);
 		String json = new String();
 		try {
